requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.30
python-dotenv==1.0.1
httpx==0.27.0
joblib==1.4.2
scikit-learn==1.4.0
selenium==4.22.0
beautifulsoup4==4.12.2
requests==2.32.3
discord.py==2.5.2
lxml==6.0.0
webdriver-manager==4.0.2
scipy==1.11.4
matplotlib==3.8.0
tensorflow-cpu==2.15.0
xgboost==2.0.3
flask==2.3.3
playwright==1.37.0